backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.0
sqlalchemy==2.0.23
PyPDF2==3.0.1
python-docx==1.1.0
jinja2==3.1.2
python-multipart==0.0.6
playwright==1.40.0